[ADD] 대결 신청 API 적용 완료

+++ domain/src/main/java/kr/co/domain/model/Keyword.kt
@@ -3,5 +3,5 @@ package kr.co.domain.model
 data class Keyword(
     val meaning: String,
     val word: String,
-    val wordId: Int
+    val wordId: Long
 )
